requirements.txt
# 基础依赖
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0

# LangChain 相关
langchain==0.2.2
langchain-cli==0.0.24
langchain-community==0.2.3
langchain-core==0.2.4
langchain-experimental==0.0.60
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langserve[all]==0.2.1
langsmith==0.1.73

# 核心依赖
pydantic==1.10.21
pydantic_core==2.18.4
protobuf==5.29.3
numpy==1.26.4
SQLAlchemy==2.0.30

# AI 和向量数据库
openai==1.31.1
chromadb==0.6.3
transformers==4.41.2
torch==2.0.0
playwright==1.41.2  # 添加 playwright 依赖

# 其他工具包
python-dotenv==1.0.1
requests==2.32.3
PyYAML==6.0.1

# 网页解析工具
beautifulsoup4==4.12.3
lxml==5.1.0  # 添加 lxml 解析器
